+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## Unreleased
+
+These are the changes that have not been released yet. The release notes have not yet been written.
+
+**REMOVED FEATURES:**
+
+- Can no longer target a different kernel release.
+
 ## v0.3.6 (2022-10-27)
 
 **FEATURE ENCHANCEMENTS:**